feat: add create json metadata for comments (#2049)
